Moves ApiCall for paintings to Hooks
